fix(oauth): move oauthResource seed helper under lib/db
Keeps the Better Auth 1.7 oauthResource bootstrap compatible with the
database client boundary guard added in #696.
